+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 * Added LibLog to .NET 4.x projects.
 * Added a DLL compiled for .NET 3.5.
+* Added a DLL compiled for .NET Standard 1.3.
 * Performance have been improved by 30%.
+* Added a StringBuilder pool in the "Specialized" namespace.
 
 ### v2.0.1 (2016-08-07) ###